fix: [CDS-23425]: Added Support for account level sdk default backoff strategy (#34377)
* [CDS-23425]: Added Support for account level sdk default backoff strategy

* [CDS-23425]: Added Support for account level sdk default backoff strategy

* [CDS-23425]: Added Support for account level sdk default backoff strategy

* [CDS-23425]: Added Support for account level sdk default backoff strategy

* [CDS-23425]: Added Support for account level sdk default backoff strategy

* [CDS-23425]: Added Support for account level sdk default backoff strategy
