+++ 400-rest/src/main/java/software/wings/sm/states/EcsStateHelper.java
@@ -131,6 +131,7 @@ import software.wings.helpers.ext.ecs.response.EcsDeployRollbackDataFetchRespons
 import software.wings.helpers.ext.ecs.response.EcsServiceDeployResponse;
 import software.wings.helpers.ext.k8s.request.K8sValuesLocation;
 import software.wings.service.impl.artifact.ArtifactCollectionUtils;
+import software.wings.service.impl.aws.manager.AwsHelperServiceManager;
 import software.wings.service.impl.aws.model.AwsEcsAllPhaseRollbackData;
 import software.wings.service.intfc.ActivityService;
 import software.wings.service.intfc.DelegateService;
@@ -590,8 +591,10 @@ public class EcsStateHelper {
       throw new InvalidArgumentsException(Pair.of("Cloud Provider", "Must be of type Aws Config"));
     }
     AwsConfig awsConfig = (AwsConfig) settingValue;
+    AwsHelperServiceManager.setAmazonClientSDKDefaultBackoffStrategyIfExists(executionContext,awsConfig);
     List<EncryptedDataDetail> encryptedDataDetails = secretManager.getEncryptionDetails(
         awsConfig, executionContext.getAppId(), executionContext.getWorkflowExecutionId());
+
 
     return EcsRunTaskDataBag.builder()
         .applicationAccountId(app.getAccountId())
@@ -641,6 +644,7 @@ public class EcsStateHelper {
       throw new InvalidArgumentsException(Pair.of("Cloud Provider", "Must be of type Aws Config"));
     }
     AwsConfig awsConfig = (AwsConfig) settingValue;
+    AwsHelperServiceManager.setAmazonClientSDKDefaultBackoffStrategyIfExists(context,awsConfig);
     List<EncryptedDataDetail> encryptedDataDetails =
         secretManager.getEncryptionDetails(awsConfig, context.getAppId(), context.getWorkflowExecutionId());
     EcsServiceSpecification serviceSpecification =
@@ -1164,6 +1168,7 @@ public class EcsStateHelper {
       throw new InvalidRequestException(format("Invalid setting value type: [%s]", settingValue.getClass().getName()));
     }
     AwsConfig awsConfig = (AwsConfig) settingValue;
+    AwsHelperServiceManager.setAmazonClientSDKDefaultBackoffStrategyIfExists(context,awsConfig);
     List<EncryptedDataDetail> encryptionDetails =
         secretManager.getEncryptionDetails(awsConfig, context.getAppId(), context.getWorkflowExecutionId());
     String region = ecsInfrastructureMapping.getRegion();
