+++ 400-rest/src/main/java/software/wings/sm/states/provision/CloudFormationRollbackStackState.java
@@ -55,6 +55,7 @@ import software.wings.sm.ExecutionContext;
 import software.wings.sm.ExecutionContextImpl;
 import software.wings.sm.ExecutionResponse;
 import software.wings.sm.StateType;
+import software.wings.service.impl.aws.manager.AwsHelperServiceManager;
 
 import com.amazonaws.services.cloudformation.model.StackStatus;
 import com.github.reinert.jjschema.SchemaIgnore;
@@ -232,7 +233,7 @@ public class CloudFormationRollbackStackState extends CloudFormationState {
        * For details, see CD-4767
        */
       AwsConfig awsConfig = getAwsConfig(configParameter.getAwsConfigId());
-      setAmazonClientSDKDefaultBackoffStrategyIfExists(context, awsConfig);
+      AwsHelperServiceManager.setAmazonClientSDKDefaultBackoffStrategyIfExists(context,awsConfig);
       String roleArnRendered = executionContext.renderExpression(configParameter.getCloudFormationRoleArn());
       CloudFormationCreateStackRequestBuilder builder = CloudFormationCreateStackRequest.builder().awsConfig(awsConfig);
       if (CLOUDFORMATION_STACK_CREATE_URL.equals(configParameter.getCreateType())) {
@@ -323,7 +324,7 @@ public class CloudFormationRollbackStackState extends CloudFormationState {
     CloudFormationRollbackInfoElement stackElement = stackElementOptional.get();
     ExecutionContextImpl executionContext = (ExecutionContextImpl) context;
     AwsConfig awsConfig = getAwsConfig(stackElement.getAwsConfigId());
-    setAmazonClientSDKDefaultBackoffStrategyIfExists(context, awsConfig);
+    AwsHelperServiceManager.setAmazonClientSDKDefaultBackoffStrategyIfExists(context,awsConfig);
     DelegateTask delegateTask;
     if (!stackElement.isStackExisted()) {
       CloudFormationDeleteStackRequest request =
